- refactor firebase into its own file
